[chore] [exporterhelper] Remove persistent queue redundant abstraction
`persistentQueue` abstraction became very thin after recent changes. This change move all the logic from `persistentContiguousStorage` to `persistentQueue`
