+++ exporter/exporterhelper/internal/persistent_queue.go
@@ -4,44 +4,376 @@
 package internal // import "go.opentelemetry.io/collector/exporter/exporterhelper/internal"
 
 import (
+	"bytes"
 	"context"
+	"encoding/binary"
 	"errors"
+	"fmt"
+	"strconv"
+	"sync"
+
+	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/exporter"
 	"go.opentelemetry.io/collector/extension/experimental/storage"
 )
 
+const (
+	zapKey           = "key"
+	zapErrorCount    = "errorCount"
+	zapNumberOfItems = "numberOfItems"
+
+	readIndexKey                = "ri"
+	writeIndexKey               = "wi"
+	currentlyDispatchedItemsKey = "di"
+)
+
 var (
+	errValueNotSet        = errors.New("value not set")
 	errNoStorageClient    = errors.New("no storage client extension found")
 	errWrongExtensionType = errors.New("requested extension is not a storage extension")
 )
 
-// persistentQueue holds the queue backed by file storage
+// persistentQueue provides a persistent queue implementation backed by file storage extension
+//
+// Write index describes the position at which next item is going to be stored.
+// Read index describes which item needs to be read next.
+// When Write index = Read index, no elements are in the queue.
+//
+// The items currently dispatched by consumers are not deleted until the processing is finished.
+// Their list is stored under a separate key.
+//
+//	┌───────file extension-backed queue───────┐
+//	│                                         │
+//	│     ┌───┐     ┌───┐ ┌───┐ ┌───┐ ┌───┐   │
+//	│ n+1 │ n │ ... │ 4 │ │ 3 │ │ 2 │ │ 1 │   │
+//	│     └───┘     └───┘ └─x─┘ └─|─┘ └─x─┘   │
+//	│                       x     |     x     │
+//	└───────────────────────x─────|─────x─────┘
+//	   ▲              ▲     x     |     x
+//	   │              │     x     |     xxxx deleted
+//	   │              │     x     |
+//	 write          read    x     └── currently dispatched item
+//	 index          index   x
+//	                        xxxx deleted
 type persistentQueue[T any] struct {
-	*persistentContiguousStorage[T]
-	set       exporter.CreateSettings
-	storageID component.ID
-	dataType  component.DataType
+	set         exporter.CreateSettings
+	storageID   component.ID
+	dataType    component.DataType
+	client      storage.Client
+	unmarshaler func(data []byte) (T, error)
+	marshaler   func(req T) ([]byte, error)
+
+	putChan  chan struct{}
+	stopChan chan struct{}
+	capacity uint64
+
+	mu                       sync.Mutex
+	readIndex                uint64
+	writeIndex               uint64
+	currentlyDispatchedItems []uint64
+	refClient                int64
 }
 
 // NewPersistentQueue creates a new queue backed by file storage; name and signal must be a unique combination that identifies the queue storage
 func NewPersistentQueue[T any](capacity int, dataType component.DataType, storageID component.ID, marshaler func(req T) ([]byte, error), unmarshaler func([]byte) (T, error), set exporter.CreateSettings) Queue[T] {
 	return &persistentQueue[T]{
-		persistentContiguousStorage: newPersistentContiguousStorage(set.Logger, uint64(capacity), marshaler, unmarshaler),
-		set:                         set,
-		storageID:                   storageID,
-		dataType:                    dataType,
+		set:         set,
+		storageID:   storageID,
+		dataType:    dataType,
+		unmarshaler: unmarshaler,
+		marshaler:   marshaler,
+		capacity:    uint64(capacity),
+		putChan:     make(chan struct{}, capacity),
+		stopChan:    make(chan struct{}),
 	}
 }
 
 // Start starts the persistentQueue with the given number of consumers.
 func (pq *persistentQueue[T]) Start(ctx context.Context, host component.Host) error {
-	storageClient, err := toStorageClient(ctx, pq.storageID, host, pq.set.ID, pq.dataType)
+	var err error
+	pq.client, err = toStorageClient(ctx, pq.storageID, host, pq.set.ID, pq.dataType)
 	if err != nil {
 		return err
 	}
-	pq.persistentContiguousStorage.start(ctx, storageClient)
+
+	pq.refClient = 1
+	pq.initPersistentContiguousStorage(ctx)
+	// Make sure the leftover requests are handled
+	pq.retrieveAndEnqueueNotDispatchedReqs(ctx)
+
+	// Ensure the communication channel has the same size as the queue
+	// We might already have items here from requeueing non-dispatched requests
+	for len(pq.putChan) < int(pq.size()) {
+		pq.putChan <- struct{}{}
+	}
+	return nil
+}
+
+func (pq *persistentQueue[T]) initPersistentContiguousStorage(ctx context.Context) {
+	riOp := storage.GetOperation(readIndexKey)
+	wiOp := storage.GetOperation(writeIndexKey)
+
+	err := pq.client.Batch(ctx, riOp, wiOp)
+	if err == nil {
+		pq.readIndex, err = bytesToItemIndex(riOp.Value)
+	}
+
+	if err == nil {
+		pq.writeIndex, err = bytesToItemIndex(wiOp.Value)
+	}
+
+	if err != nil {
+		if errors.Is(err, errValueNotSet) {
+			pq.set.Logger.Info("Initializing new persistent queue")
+		} else {
+			pq.set.Logger.Error("Failed getting read/write index, starting with new ones", zap.Error(err))
+		}
+		pq.readIndex = 0
+		pq.writeIndex = 0
+	}
+}
+
+// Poll returns the next available item from the queue, or blocks until one is available.
+// If the queue is stopped, returns (QueueRequest{}, false)
+func (pq *persistentQueue[T]) Poll() (QueueRequest[T], bool) {
+	for {
+		select {
+		case <-pq.stopChan:
+			return QueueRequest[T]{}, false
+		case <-pq.putChan:
+			req, found := pq.getNextItem(context.Background())
+			if found {
+				return req, true
+			}
+		}
+	}
+}
+
+func (pq *persistentQueue[T]) size() uint64 {
+	return pq.writeIndex - pq.readIndex
+}
+
+// Size returns the number of currently available items, which were not picked by consumers yet
+func (pq *persistentQueue[T]) Size() int {
+	pq.mu.Lock()
+	defer pq.mu.Unlock()
+	return int(pq.size())
+}
+
+// Capacity returns the number of currently available items, which were not picked by consumers yet
+func (pq *persistentQueue[T]) Capacity() int {
+	return int(pq.capacity)
+}
+
+func (pq *persistentQueue[T]) Shutdown(ctx context.Context) error {
+	close(pq.stopChan)
+	// Hold the lock only for `refClient`.
+	pq.mu.Lock()
+	defer pq.mu.Unlock()
+	return pq.unrefClient(ctx)
+}
+
+// unrefClient unrefs the client, and closes if no more references. Callers MUST hold the mutex.
+// This is needed because consumers of the queue may still process the requests while the queue is shutting down or immediately after.
+func (pq *persistentQueue[T]) unrefClient(ctx context.Context) error {
+	pq.refClient--
+	if pq.refClient == 0 {
+		return pq.client.Close(ctx)
+	}
+	return nil
+}
+
+// Offer inserts the specified element into this queue if it is possible to do so immediately
+// without violating capacity restrictions. If success returns no error.
+// It returns ErrQueueIsFull if no space is currently available.
+func (pq *persistentQueue[T]) Offer(ctx context.Context, req T) error {
+	pq.mu.Lock()
+	defer pq.mu.Unlock()
+	return pq.putInternal(ctx, req)
+}
+
+// putInternal is the internal version that requires caller to hold the mutex lock.
+func (pq *persistentQueue[T]) putInternal(ctx context.Context, req T) error {
+	if pq.size() >= pq.capacity {
+		pq.set.Logger.Warn("Maximum queue capacity reached")
+		return ErrQueueIsFull
+	}
+
+	itemKey := getItemKey(pq.writeIndex)
+	pq.writeIndex++
+
+	reqBuf, err := pq.marshaler(req)
+	if err != nil {
+		return err
+	}
+	err = pq.client.Batch(ctx,
+		storage.SetOperation(writeIndexKey, itemIndexToBytes(pq.writeIndex)),
+		storage.SetOperation(itemKey, reqBuf))
+
+	// Inform the loop that there's some data to process
+	pq.putChan <- struct{}{}
+
+	return err
+}
+
+// getNextItem pulls the next available item from the persistent storage; if none is found, returns (nil, false)
+func (pq *persistentQueue[T]) getNextItem(ctx context.Context) (QueueRequest[T], bool) {
+	pq.mu.Lock()
+	defer pq.mu.Unlock()
+
+	// If called in the same time with Shutdown, make sure client is not closed.
+	if pq.refClient <= 0 {
+		return QueueRequest[T]{}, false
+	}
+
+	if pq.readIndex == pq.writeIndex {
+		return QueueRequest[T]{}, false
+	}
+	index := pq.readIndex
+	// Increase here, so even if errors happen below, it always iterates
+	pq.readIndex++
+
+	pq.currentlyDispatchedItems = append(pq.currentlyDispatchedItems, index)
+	getOp := storage.GetOperation(getItemKey(index))
+	err := pq.client.Batch(ctx,
+		storage.SetOperation(readIndexKey, itemIndexToBytes(pq.readIndex)),
+		storage.SetOperation(currentlyDispatchedItemsKey, itemIndexArrayToBytes(pq.currentlyDispatchedItems)),
+		getOp)
+
+	var request T
+	if err == nil {
+		request, err = pq.unmarshaler(getOp.Value)
+	}
+
+	if err != nil {
+		pq.set.Logger.Debug("Failed to dispatch item", zap.Error(err))
+		// We need to make sure that currently dispatched items list is cleaned
+		if err = pq.itemDispatchingFinish(ctx, index); err != nil {
+			pq.set.Logger.Error("Error deleting item from queue", zap.Error(err))
+		}
+
+		return QueueRequest[T]{}, false
+	}
+
+	req := newQueueRequest[T](context.Background(), request)
+	// If all went well so far, cleanup will be handled by callback
+	pq.refClient++
+	req.onProcessingFinishedFunc = func() {
+		pq.mu.Lock()
+		defer pq.mu.Unlock()
+		if err = pq.itemDispatchingFinish(ctx, index); err != nil {
+			pq.set.Logger.Error("Error deleting item from queue", zap.Error(err))
+		}
+		if err = pq.unrefClient(ctx); err != nil {
+			pq.set.Logger.Error("Error closing the storage client", zap.Error(err))
+		}
+	}
+	return req, true
+}
+
+// retrieveAndEnqueueNotDispatchedReqs gets the items for which sending was not finished, cleans the storage
+// and moves the items at the back of the queue.
+func (pq *persistentQueue[T]) retrieveAndEnqueueNotDispatchedReqs(ctx context.Context) {
+	var dispatchedItems []uint64
+
+	pq.mu.Lock()
+	defer pq.mu.Unlock()
+	pq.set.Logger.Debug("Checking if there are items left for dispatch by consumers")
+	itemKeysBuf, err := pq.client.Get(ctx, currentlyDispatchedItemsKey)
+	if err == nil {
+		dispatchedItems, err = bytesToItemIndexArray(itemKeysBuf)
+	}
+	if err != nil {
+		pq.set.Logger.Error("Could not fetch items left for dispatch by consumers", zap.Error(err))
+		return
+	}
+
+	if len(dispatchedItems) == 0 {
+		pq.set.Logger.Debug("No items left for dispatch by consumers")
+		return
+	}
+
+	pq.set.Logger.Info("Fetching items left for dispatch by consumers", zap.Int(zapNumberOfItems, len(dispatchedItems)))
+	retrieveBatch := make([]storage.Operation, len(dispatchedItems))
+	cleanupBatch := make([]storage.Operation, len(dispatchedItems))
+	for i, it := range dispatchedItems {
+		key := getItemKey(it)
+		retrieveBatch[i] = storage.GetOperation(key)
+		cleanupBatch[i] = storage.DeleteOperation(key)
+	}
+	retrieveErr := pq.client.Batch(ctx, retrieveBatch...)
+	cleanupErr := pq.client.Batch(ctx, cleanupBatch...)
+
+	if cleanupErr != nil {
+		pq.set.Logger.Debug("Failed cleaning items left by consumers", zap.Error(cleanupErr))
+	}
+
+	if retrieveErr != nil {
+		pq.set.Logger.Warn("Failed retrieving items left by consumers", zap.Error(retrieveErr))
+		return
+	}
+
+	errCount := 0
+	for _, op := range retrieveBatch {
+		if op.Value == nil {
+			pq.set.Logger.Warn("Failed retrieving item", zap.String(zapKey, op.Key), zap.Error(errValueNotSet))
+			continue
+		}
+		req, err := pq.unmarshaler(op.Value)
+		// If error happened or item is nil, it will be efficiently ignored
+		if err != nil {
+			pq.set.Logger.Warn("Failed unmarshalling item", zap.String(zapKey, op.Key), zap.Error(err))
+			continue
+		}
+		if pq.putInternal(ctx, req) != nil {
+			errCount++
+		}
+	}
+
+	if errCount > 0 {
+		pq.set.Logger.Error("Errors occurred while moving items for dispatching back to queue",
+			zap.Int(zapNumberOfItems, len(retrieveBatch)), zap.Int(zapErrorCount, errCount))
+	} else {
+		pq.set.Logger.Info("Moved items for dispatching back to queue",
+			zap.Int(zapNumberOfItems, len(retrieveBatch)))
+	}
+}
+
+// itemDispatchingFinish removes the item from the list of currently dispatched items and deletes it from the persistent queue
+func (pq *persistentQueue[T]) itemDispatchingFinish(ctx context.Context, index uint64) error {
+	lenCDI := len(pq.currentlyDispatchedItems)
+	for i := 0; i < lenCDI; i++ {
+		if pq.currentlyDispatchedItems[i] == index {
+			pq.currentlyDispatchedItems[i] = pq.currentlyDispatchedItems[lenCDI-1]
+			pq.currentlyDispatchedItems = pq.currentlyDispatchedItems[:lenCDI-1]
+			break
+		}
+	}
+
+	setOp := storage.SetOperation(currentlyDispatchedItemsKey, itemIndexArrayToBytes(pq.currentlyDispatchedItems))
+	deleteOp := storage.DeleteOperation(getItemKey(index))
+	if err := pq.client.Batch(ctx, setOp, deleteOp); err != nil {
+		// got an error, try to gracefully handle it
+		pq.set.Logger.Warn("Failed updating currently dispatched items, trying to delete the item first",
+			zap.Error(err))
+	} else {
+		// Everything ok, exit
+		return nil
+	}
+
+	if err := pq.client.Batch(ctx, deleteOp); err != nil {
+		// Return an error here, as this indicates an issue with the underlying storage medium
+		return fmt.Errorf("failed deleting item from queue, got error from storage: %w", err)
+	}
+
+	if err := pq.client.Batch(ctx, setOp); err != nil {
+		// even if this fails, we still have the right dispatched items in memory
+		// at worst, we'll have the wrong list in storage, and we'll discard the nonexistent items during startup
+		return fmt.Errorf("failed updating currently dispatched items, but deleted item successfully: %w", err)
+	}
+
 	return nil
 }
 
@@ -57,4 +389,46 @@ func toStorageClient(ctx context.Context, storageID component.ID, host component
 	}
 
 	return storageExt.GetClient(ctx, component.KindExporter, ownerID, string(signal))
+}
+
+func getItemKey(index uint64) string {
+	return strconv.FormatUint(index, 10)
+}
+
+func itemIndexToBytes(value uint64) []byte {
+	return binary.LittleEndian.AppendUint64([]byte{}, value)
+}
+
+func bytesToItemIndex(b []byte) (uint64, error) {
+	val := uint64(0)
+	if b == nil {
+		return val, errValueNotSet
+	}
+	err := binary.Read(bytes.NewReader(b), binary.LittleEndian, &val)
+	return val, err
+}
+
+func itemIndexArrayToBytes(arr []uint64) []byte {
+	size := len(arr)
+	buf := make([]byte, 0, 4+size*8)
+	buf = binary.LittleEndian.AppendUint32(buf, uint32(size))
+	for _, item := range arr {
+		buf = binary.LittleEndian.AppendUint64(buf, item)
+	}
+	return buf
+}
+
+func bytesToItemIndexArray(b []byte) ([]uint64, error) {
+	if len(b) == 0 {
+		return nil, nil
+	}
+	var size uint32
+	reader := bytes.NewReader(b)
+	if err := binary.Read(reader, binary.LittleEndian, &size); err != nil {
+		return nil, err
+	}
+
+	val := make([]uint64, size)
+	err := binary.Read(reader, binary.LittleEndian, &val)
+	return val, err
 }
